refactor(parquet): simplify source worker boundary
